cloud_function_script/report_error/requirements.txt
requests==2.22.0
opencv-python==4.1.0.25
Flask==2.0.1
google-cloud-bigquery